// src/app/data-access/models/device.model.ts
export enum DeviceStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  ALL = "ALL",
}

//Device Network
export enum DeviceNetwork {
  lor = "lor",
  N01 = "N01",
  R234 = "R234",
  sig = "sig",
  uncl = "uncl",
  ALL = "ALL",
}

/**
 * Device model
 */
export interface Device {
  deviceId: string;
  deviceName: string;
  assetId: string;
  assetName?: string;
  companyId?: string;
  companyName?: string;
  deviceCategName?: string;
  deviceCategoryName?: string;
  locationId?: String;
  locationName?: string;
  statusId: String;
  statusName?: DeviceStatus;
  locationAddress1?: String;
  locationAddress2: String;
  manufDeviceId?: string;
  networkId: string;
  networkName: string;
}
// ...
export interface DeviceCategory
  extends DeviceCategoryDirectory,
    DeviceCategoryView {}

export interface DeviceCategoryDirectory {
  deviceCategId: string;
  deviceCategName: string;
  companyName: string;
  companyId?: string;
  networkId: string;
  networkName: string;
}

export interface DeviceCategoryView {
  deviceCategName: string;
  deviceCategDesc: string;
  companyId?: string;
  companyName: string;
}

export interface DeviceFormData {
  companyId: string;
  userId: string;
  deviceAdd?: DeviceDetail[];
  deviceDel?: { deviceId: string }[];
}

export interface DeviceCategoryFormData {
  companyId: string;
  userId: string;
  clientDeviceCategId: string;
  deviceCategName: string;
  deviceCategDesc: string;
  deviceAdd?: { deviceId: string }[];
  deviceDel?: { deviceId: string }[];
}